+++ player_achievements.py
@@ -126,7 +126,7 @@ def add_args_fetch(parser: ArgumentParser, config: Optional[ConfigParser] = None
 							help='Fetch stats for accounts that have been active since RELEASE/DAYS')
 		parser.add_argument('--inactive-since', type=str,  default=None, metavar='RELEASE',
 							help='Fetch stats for accounts that have been inactive since RELEASE/DAYS')	
-		parser.add_argument('--cache_valid', type=int, default=0, metavar='DAYS',
+		parser.add_argument('--cache_valid', type=float, default=0, metavar='DAYS',
 							help='Fetch only accounts with stats older than DAYS')		
 		parser.add_argument('--distributed', '--dist',type=str, dest='distributed', metavar='I:N', 
 							default=None, help='Distributed fetching for accounts: id %% N == I')
